Replace old implementation with new
