--- models/demos/vanilla_unet_new/demo/demo_utils.py
@@ -1,388 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-import os
-import random
-
-import numpy as np
-import torch
-from loguru import logger
-from medpy.filter.binary import largest_connected_component
-from skimage.exposure import rescale_intensity
-from skimage.io import imread
-from skimage.transform import resize
-from torch.utils.data import Dataset
-
-
-def data_loader(args):
-    dataset = BrainSegmentationDataset(
-        image_path=args.image,  # single image path
-        mask_path=args.mask,  # single mask path
-        image_size=args.image_size,
-        transform=None,  # Add any transformations if necessary
-    )
-
-    loader = torch.utils.data.DataLoader(dataset, batch_size=args.batch_size, drop_last=False, num_workers=0)
-    return loader
-
-
-def data_loader_imageset(args):
-    dataset = BrainSegmentationDatasetImageset(
-        images_dir=args.images,
-        subset="validation",
-        image_size=args.image_size,
-        random_sampling=False,
-    )
-    loader = torch.utils.data.DataLoader(dataset, batch_size=args.batch_size, drop_last=False, num_workers=0)
-    return loader
-
-
-def crop_sample(x):
-    volume, mask = x
-    volume[volume < np.max(volume) * 0.1] = 0
-    z_projection = np.max(np.max(np.max(volume, axis=-1), axis=-1), axis=-1)
-    z_nonzero = np.nonzero(z_projection)
-    z_min = np.min(z_nonzero)
-    z_max = np.max(z_nonzero) + 1
-    y_projection = np.max(np.max(np.max(volume, axis=0), axis=-1), axis=-1)
-    y_nonzero = np.nonzero(y_projection)
-    y_min = np.min(y_nonzero)
-    y_max = np.max(y_nonzero) + 1
-    x_projection = np.max(np.max(np.max(volume, axis=0), axis=0), axis=-1)
-    x_nonzero = np.nonzero(x_projection)
-    x_min = np.min(x_nonzero)
-    x_max = np.max(x_nonzero) + 1
-    return (
-        volume[z_min:z_max, y_min:y_max, x_min:x_max],
-        mask[z_min:z_max, y_min:y_max, x_min:x_max],
-    )
-
-
-def pad_sample(x):
-    volume, mask = x
-    a = volume.shape[1]
-    b = volume.shape[2]
-    if a == b:
-        return volume, mask
-    diff = (max(a, b) - min(a, b)) / 2.0
-    if a > b:
-        padding = ((0, 0), (0, 0), (int(np.floor(diff)), int(np.ceil(diff))))
-    else:
-        padding = ((0, 0), (int(np.floor(diff)), int(np.ceil(diff))), (0, 0))
-    mask = np.pad(mask, padding, mode="constant", constant_values=0)
-    padding = padding + ((0, 0),)
-    volume = np.pad(volume, padding, mode="constant", constant_values=0)
-    return volume, mask
-
-
-def resize_sample(x, size=256):
-    volume, mask = x
-    v_shape = volume.shape
-    out_shape = (1, size[0], size[1])
-    mask = resize(
-        mask,
-        output_shape=out_shape,
-        order=0,
-        mode="constant",
-        cval=0,
-        anti_aliasing=False,
-    )
-    out_shape = out_shape + (3,)
-    volume = resize(
-        volume,
-        output_shape=out_shape,
-        order=2,
-        mode="constant",
-        cval=0,
-        anti_aliasing=False,
-    )
-    return volume, mask
-
-
-def normalize_volume(volume):
-    p10 = np.percentile(volume, 10)
-    p99 = np.percentile(volume, 99)
-    volume = rescale_intensity(volume, in_range=(p10, p99))
-    m = np.mean(volume, axis=(0, 1, 2))
-    s = np.std(volume, axis=(0, 1, 2))
-    volume = (volume - m) / s
-    return volume
-
-
-def crop_sample_imageset(x):
-    volume, mask = x
-    volume[volume < np.max(volume) * 0.1] = 0
-    z_projection = np.max(np.max(np.max(volume, axis=-1), axis=-1), axis=-1)
-    z_nonzero = np.nonzero(z_projection)
-    z_min = np.min(z_nonzero)
-    z_max = np.max(z_nonzero) + 1
-    y_projection = np.max(np.max(np.max(volume, axis=0), axis=-1), axis=-1)
-    y_nonzero = np.nonzero(y_projection)
-    y_min = np.min(y_nonzero)
-    y_max = np.max(y_nonzero) + 1
-    x_projection = np.max(np.max(np.max(volume, axis=0), axis=0), axis=-1)
-    x_nonzero = np.nonzero(x_projection)
-    x_min = np.min(x_nonzero)
-    x_max = np.max(x_nonzero) + 1
-    return (
-        volume[z_min:z_max, y_min:y_max, x_min:x_max],
-        mask[z_min:z_max, y_min:y_max, x_min:x_max],
-    )
-
-
-def pad_sample_imageset(x):
-    volume, mask = x
-    a = volume.shape[1]
-    b = volume.shape[2]
-    if a == b:
-        return volume, mask
-    diff = (max(a, b) - min(a, b)) / 2.0
-    if a > b:
-        padding = ((0, 0), (0, 0), (int(np.floor(diff)), int(np.ceil(diff))))
-    else:
-        padding = ((0, 0), (int(np.floor(diff)), int(np.ceil(diff))), (0, 0))
-    mask = np.pad(mask, padding, mode="constant", constant_values=0)
-    padding = padding + ((0, 0),)
-    volume = np.pad(volume, padding, mode="constant", constant_values=0)
-    return volume, mask
-
-
-def resize_sample_imageset(x, size=256, resolution=(480, 640)):
-    volume, mask = x
-    v_shape = volume.shape
-    out_shape = (v_shape[0], resolution[0], resolution[1])
-    mask = resize(
-        mask,
-        output_shape=out_shape,
-        order=0,
-        mode="constant",
-        cval=0,
-        anti_aliasing=False,
-    )
-    out_shape = out_shape + (v_shape[3],)
-    volume = resize(
-        volume,
-        output_shape=out_shape,
-        order=2,
-        mode="constant",
-        cval=0,
-        anti_aliasing=False,
-    )
-    return volume, mask
-
-
-def normalize_volume_imageset(volume):
-    p10 = np.percentile(volume, 10)
-    p99 = np.percentile(volume, 99)
-    volume = rescale_intensity(volume, in_range=(p10, p99))
-    m = np.mean(volume, axis=(0, 1, 2))
-    s = np.std(volume, axis=(0, 1, 2))
-    volume = (volume - m) / s
-    return volume
-
-
-class BrainSegmentationDataset(Dataset):
-    """Brain MRI dataset for FLAIR abnormality segmentation (single image with single mask)"""
-
-    in_channels = 3
-    out_channels = 1
-
-    def __init__(
-        self,
-        image_path,
-        mask_path,
-        transform=None,
-        image_size=(480, 640),
-    ):
-        """
-        Args:
-            image_path (str): Path to the input image.
-            mask_path (str): Path to the mask image.
-            transform (callable, optional): Optional transform to be applied on an image and its mask.
-            image_size (tuple, optional): Size to resize the image and mask to.
-        """
-        # Load image and mask
-        self.image = imread(image_path)
-        self.mask = imread(mask_path, as_gray=True)
-        self.image = np.expand_dims(self.image, axis=0)
-        self.mask = np.expand_dims(self.mask, axis=0)
-
-        # Preprocessing
-        self.image, self.mask = crop_sample((self.image, self.mask))
-        self.image, self.mask = pad_sample((self.image, self.mask))
-        self.image, self.mask = resize_sample((self.image, self.mask), size=image_size)
-        self.image = normalize_volume(self.image)
-
-        # Add channel dimension for image and mask
-        self.image = self.image.transpose(0, 3, 1, 2)  # Change from (H, W, C) to (C, H, W)
-        # Optional transformation
-        self.transform = transform
-
-    def __len__(self):
-        return 1  # Only one image with one mask
-
-    def __getitem__(self, idx):
-        # Directly return the image and mask as tensors
-        image_tensor = torch.from_numpy(self.image.astype(np.float32))
-        mask_tensor = torch.from_numpy(self.mask.astype(np.float32))
-
-        if self.transform:
-            image_tensor, mask_tensor = self.transform((image_tensor, mask_tensor))
-
-        return image_tensor, mask_tensor
-
-
-class BrainSegmentationDatasetImageset(Dataset):
-    """Brain MRI dataset for FLAIR abnormality segmentation"""
-
-    in_channels = 3
-    out_channels = 1
-
-    def __init__(
-        self,
-        images_dir,
-        transform=None,
-        image_size=256,
-        subset="train",
-        random_sampling=True,
-        validation_cases=1,
-        seed=42,
-    ):
-        assert subset in ["all", "train", "validation"]
-
-        # read images
-        volumes = {}
-        masks = {}
-        logger.info("reading {} images...".format(subset))
-        for dirpath, dirnames, filenames in os.walk(images_dir):
-            image_slices = []
-            mask_slices = []
-            for filename in sorted(
-                filter(lambda f: ".tif" in f, filenames),
-                key=lambda x: int(x.split(".")[-2].split("_")[4]),
-            ):
-                filepath = os.path.join(dirpath, filename)
-                if "mask" in filename:
-                    mask_slices.append(imread(filepath, as_gray=True))
-                else:
-                    image_slices.append(imread(filepath))
-            if len(image_slices) > 0:
-                patient_id = dirpath.split("/")[-1]
-                volumes[patient_id] = np.array(image_slices[1:-1])
-                masks[patient_id] = np.array(mask_slices[1:-1])
-
-        self.patients = sorted(volumes)
-
-        # select cases to subset
-        if not subset == "all":
-            random.seed(seed)
-            validation_patients = random.sample(self.patients, k=validation_cases)
-            if subset == "validation":
-                self.patients = validation_patients
-            else:
-                self.patients = sorted(list(set(self.patients).difference(validation_patients)))
-
-        logger.info("preprocessing {} volumes...".format(subset))
-        # create list of tuples (volume, mask)
-        self.volumes = [(volumes[k], masks[k]) for k in self.patients]
-
-        logger.info("cropping {} volumes...".format(subset))
-        # crop to smallest enclosing volume
-        self.volumes = [crop_sample_imageset(v) for v in self.volumes]
-
-        logger.info("padding {} volumes...".format(subset))
-        # pad to square
-        self.volumes = [pad_sample_imageset(v) for v in self.volumes]
-
-        logger.info("resizing {} volumes...".format(subset))
-        # resize
-        self.volumes = [resize_sample_imageset(v, size=image_size) for v in self.volumes]
-
-        logger.info("normalizing {} volumes...".format(subset))
-        # normalize channel-wise
-        self.volumes = [(normalize_volume_imageset(v), m) for v, m in self.volumes]
-
-        # probabilities for sampling slices based on masks
-        self.slice_weights = [m.sum(axis=-1).sum(axis=-1) for v, m in self.volumes]
-        self.slice_weights = [(s + (s.sum() * 0.1 / len(s))) / (s.sum() * 1.1) for s in self.slice_weights]
-
-        # add channel dimension to masks
-        self.volumes = [(v, m[..., np.newaxis]) for (v, m) in self.volumes]
-
-        logger.info("done creating {} dataset".format(subset))
-
-        # create global index for patient and slice (idx -> (p_idx, s_idx))
-        num_slices = [v.shape[0] for v, m in self.volumes]
-        self.patient_slice_index = list(
-            zip(
-                sum([[i] * num_slices[i] for i in range(len(num_slices))], []),
-                sum([list(range(x)) for x in num_slices], []),
-            )
-        )
-
-        self.random_sampling = random_sampling
-
-        self.transform = transform
-
-    def __len__(self):
-        return len(self.patient_slice_index)
-
-    def __getitem__(self, idx):
-        patient = self.patient_slice_index[idx][0]
-        slice_n = self.patient_slice_index[idx][1]
-
-        if self.random_sampling:
-            patient = np.random.randint(len(self.volumes))
-            slice_n = np.random.choice(range(self.volumes[patient][0].shape[0]), p=self.slice_weights[patient])
-
-        v, m = self.volumes[patient]
-        image = v[slice_n]
-        mask = m[slice_n]
-
-        if self.transform is not None:
-            image, mask = self.transform((image, mask))
-
-        # fix dimensions (C, H, W)
-        image = image.transpose(2, 0, 1)
-        mask = mask.transpose(2, 0, 1)
-
-        image_tensor = torch.from_numpy(image.astype(np.float32))
-        mask_tensor = torch.from_numpy(mask.astype(np.float32))
-
-        # return tensors
-        return image_tensor, mask_tensor
-
-
-def postprocess_per_volume(input_list, pred_list, true_list, patient_slice_index, patients):
-    volumes = {}
-    num_slices = np.bincount([p[0] for p in patient_slice_index])
-    index = 0
-    for p in range(len(num_slices)):
-        volume_in = np.array(input_list[index : index + num_slices[p]])
-        volume_pred = np.round(np.array(pred_list[index : index + num_slices[p]])).astype(int)
-        volume_pred = largest_connected_component(volume_pred)
-        volume_true = np.array(true_list[index : index + num_slices[p]])
-        volumes[patients[p]] = (volume_in, volume_pred, volume_true)
-        index += num_slices[p]
-    return volumes
-
-
-def gray2rgb(image):
-    w, h = image.shape
-    image += np.abs(np.min(image))
-    image_max = np.abs(np.max(image))
-    if image_max > 0:
-        image /= image_max
-    ret = np.empty((w, h, 3), dtype=np.uint8)
-    ret[:, :, 2] = ret[:, :, 1] = ret[:, :, 0] = image * 255
-    return ret
-
-
-def outline(image, mask, color):
-    mask = np.round(mask)
-    yy, xx = np.nonzero(mask)
-    for y, x in zip(yy, xx):
-        if 0.0 < np.mean(mask[max(0, y - 1) : y + 2, max(0, x - 1) : x + 2]) < 1.0:
-            image[max(0, y) : y + 1, max(0, x) : x + 1] = color
-    return image
